rename reviewActions -> mediaActions and reviewConstants -> mediaConstants
